Adding more detailed documentation in the README (#52)
* Adding documentation about host, port and namespace

* MARQUEZ_HOST, MARQUEZ_PORT, and MARQUEZ_NAMESPACE are valuable env vars to set

Signed-off-by: Bob Briski <rbriski@gmail.com>

* Additional documentation around extractors

This is pretty tough to do with any kind of precision right now.  It's clear that the underlying API is in flux as new objects to handle metadata are being added.  This is meant as a placeholder right now to let people know about the fact that extractors exist so they don't spend a bunch of time trying to figure out why it doesn't work out of the box

Signed-off-by: Bob Briski <rbriski@gmail.com>

* Removed MARQUEZ_HOST and MARQUEZ_PORT and replaced with MARQUEZ_URL

A note, in the marquez_client code, the DEFAULT_MARQUEZ_URL is http://localhost:8080.  The port at 8080 seems odd since the default port on the marquez server is actually 5000.  Not pertinent to this PR though.

Signed-off-by: Bob Briski <rbriski@gmail.com>

Co-authored-by: Willy Lulciuc <willy@datakin.com>